# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "实验室人脸识别打卡系统",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/admin/statistics": {
      "get": {
        "tags": [
          "管理员"
        ],
        "summary": "Statistics",
        "operationId": "statistics_api_v1_admin_statistics_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 90,
              "minimum": 7,
              "default": 14,
              "title": "Days"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminStatistics"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/attendance-sessions": {
      "get": {
        "tags": [
          "管理员"
        ],
        "summary": "List Attendance",
        "operationId": "list_attendance_api_v1_admin_attendance_sessions_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "User Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/AttendanceStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "start_at",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start At"
            }
          },
          {
            "name": "end_at",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End At"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_AttendancePublic_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/attendance-sessions/{session_id}/correct": {
      "post": {
        "tags": [
          "管理员"
        ],
        "summary": "Correct Attendance",
        "operationId": "correct_attendance_api_v1_admin_attendance_sessions__session_id__correct_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AttendanceCorrection"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttendancePublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/attendance-sessions/{session_id}/invalidate": {
      "post": {
        "tags": [
          "管理员"
        ],
        "summary": "Invalidate Attendance",
        "operationId": "invalidate_attendance_api_v1_admin_attendance_sessions__session_id__invalidate_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AttendanceInvalidate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/attendance-export.csv": {
      "get": {
        "tags": [
          "管理员"
        ],
        "summary": "Export Attendance",
        "operationId": "export_attendance_api_v1_admin_attendance_export_csv_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "start_at",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start At"
            }
          },
          {
            "name": "end_at",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End At"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/devices": {
      "get": {
        "tags": [
          "管理员"
        ],
        "summary": "List Devices",
        "operationId": "list_devices_api_v1_admin_devices_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DevicePublic"
                  },
                  "type": "array",
                  "title": "Response List Devices Api V1 Admin Devices Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "管理员"
        ],
        "summary": "Create Device",
        "operationId": "create_device_api_v1_admin_devices_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceCreated"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/devices/bootstrap-local": {
      "post": {
        "tags": [
          "管理员"
        ],
        "summary": "Bootstrap Local Device",
        "description": "Create or repair the kiosk identity for the current browser installation.",
        "operationId": "bootstrap_local_device_api_v1_admin_devices_bootstrap_local_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LocalDeviceBootstrap"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceCreated"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/devices/{device_id}": {
      "patch": {
        "tags": [
          "管理员"
        ],
        "summary": "Update Device",
        "operationId": "update_device_api_v1_admin_devices__device_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Device Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DevicePublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/audit-logs": {
      "get": {
        "tags": [
          "管理员"
        ],
        "summary": "List Audit Logs",
        "operationId": "list_audit_logs_api_v1_admin_audit_logs_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "action",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Action"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_AuditLogPublic_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AdminStatistics": {
        "properties": {
          "current_count": {
            "type": "integer",
            "title": "Current Count"
          },
          "current_users": {
            "items": {
              "$ref": "#/components/schemas/CurrentUserItem"
            },
            "type": "array",
            "title": "Current Users"
          },
          "today_checkins": {
            "type": "integer",
            "title": "Today Checkins"
          },
          "today_checkouts": {
            "type": "integer",
            "title": "Today Checkouts"
          },
          "exception_count": {
            "type": "integer",
            "title": "Exception Count"
          },
          "daily": {
            "items": {
              "$ref": "#/components/schemas/DailyDuration"
            },
            "type": "array",
            "title": "Daily"
          },
          "ranking": {
            "items": {
              "$ref": "#/components/schemas/RankingItem"
            },
            "type": "array",
            "title": "Ranking"
          },
          "hourly": {
            "items": {
              "$ref": "#/components/schemas/HourlyCount"
            },
            "type": "array",
            "title": "Hourly"
          }
        },
        "type": "object",
        "required": [
          "current_count",
          "current_users",
          "today_checkins",
          "today_checkouts",
          "exception_count",
          "daily",
          "ranking",
          "hourly"
        ],
        "title": "AdminStatistics"
      },
...
      "AttendanceCorrection": {
        "properties": {
          "check_out_at": {
            "type": "string",
            "format": "date-time",
            "title": "Check Out At"
          },
          "reason": {
            "type": "string",
            "maxLength": 500,
            "minLength": 2,
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "check_out_at",
          "reason"
        ],
        "title": "AttendanceCorrection"
      },
      "AttendanceInvalidate": {
        "properties": {
          "reason": {
            "type": "string",
            "maxLength": 500,
            "minLength": 2,
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "reason"
        ],
        "title": "AttendanceInvalidate"
      },
      "AttendancePublic": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "check_in_at": {
            "type": "string",
            "format": "date-time",
            "title": "Check In At"
          },
          "check_out_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Check Out At"
          },
          "duration_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Seconds"
          },
          "status": {
            "$ref": "#/components/schemas/AttendanceStatus"
          },
          "corrected": {
            "type": "boolean",
            "title": "Corrected"
          },
          "correction_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Correction Reason"
          },
          "user_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Name"
          },
          "username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Username"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "check_in_at",
          "check_out_at",
          "duration_seconds",
          "status",
          "corrected",
          "correction_reason"
        ],
        "title": "AttendancePublic"
      },
      "AttendanceStatus": {
        "type": "string",
        "enum": [
          "OPEN",
          "CLOSED",
          "MISSING_CHECKOUT",
          "INVALID"
        ],
        "title": "AttendanceStatus"
      },
...
      "DeviceCreate": {
        "properties": {
          "code": {
            "type": "string",
            "maxLength": 32,
            "minLength": 3,
            "pattern": "^[A-Z0-9-]+$",
            "title": "Code"
          },
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name"
          },
          "location": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Location"
          }
        },
        "type": "object",
        "required": [
          "code",
          "name",
          "location"
        ],
        "title": "DeviceCreate"
      },
...
      "DevicePublic": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "location": {
            "type": "string",
            "title": "Location"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "last_seen_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Seen At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "code",
          "name",
          "location",
          "is_active",
          "last_seen_at",
          "created_at"
        ],
        "title": "DevicePublic"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Message": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "Message"
      },
      "Page_AttendancePublic_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AttendancePublic"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "maximum": 200.0,
            "minimum": 1.0,
            "title": "Page Size"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "Page[AttendancePublic]"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}